requirements.txt
flask
anthropic
google-auth
google-auth-oauthlib
google-api-python-client
schedule
python-dotenv
jinja2
requests
